feat(server): workspace migration flat entity maps service
